editor/libeditor/html/tests/test_bug599322.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=599322.patch
-->
<head>
  <title>Test for Bug 599322.patch</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=599322.patch">Mozilla Bug 599322.patch</a>
<p id="display"></p>
<div id="content">
<div id="src">src<img src="/tests/editor/libeditor/html/tests/green.png"></div>
<iframe id="dst" src="javascript:;"></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 599322.patch **/

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var src = document.getElementById("src");
  var dst = document.getElementById("dst");
  var doc = dst.contentDocument;
  doc.open();
  doc.write("<html><head><base href='http://mochi.test:8888/'></head><body></body></html>");
  doc.close();
  SimpleTest.waitForFocus(function() {
    getSelection().selectAllChildren(src);
    SimpleTest.waitForClipboard("src",
      function() {
        synthesizeKey("c", {accelKey: true});
      },
      function() {
        dst.contentDocument.designMode = "on";
        dst.focus();
        dst.contentDocument.body.focus();
        synthesizeKey("v", {accelKey: true});
        is(dst.contentDocument.querySelector("img").src,
           document.querySelector("img").src,
           "The source should be correctly set based on the base URI");
        SimpleTest.finish();
      },
      function() {
        SimpleTest.finish();
      }
    );
  });
});

</script>
</pre>
</body>
</html>
